requirements.txt
streamlit
pandas
pdfplumber
openpyxl
google-generativeai
python-dotenv
